remark周りのリファクタリング
