+++ src/store/interComponentChannelStore.tsx
@@ -11,7 +11,6 @@ import {
     useState,
 } from "react"
 import { CEventsPropsType } from "./cEventsStore"
-import { Position } from "unist"
 import { __debugPrint__ } from "../debugtool/debugtool"
 
 /*
